2일차-2

+++ java/calculate/src/kr/co/yangdoll/sungjuk/SungjukCalc.java
@@ -1,7 +1,7 @@
 package kr.co.yangdoll.sungjuk;
 
 public class SungjukCalc {
-	public int avgCalc(int kor, int math, int eng) {
+	public double avgCalc(int kor, int math, int eng) {
 		return (kor+math+eng)/3;
 	}
 	public int totalCalc(int kor, int math, int eng) {
